Pull implementations out of test file
